Add files via upload
This is a driver drowsiness detection system that has machine learning. Here I have used MRL Eye-Dataset which has 84,898 images of 37 different persons (33 men and 4 women). So, basically, this system alerts the driver when he feels drowsy and it will generate an alarm to reduce the chance of accidents. I have used Jupyter notebook for coding.